// src_frontend/api-client/models/ProductTypeAccordionInBestellWizard.ts
/* tslint:disable */
/* eslint-disable */
/**
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface ProductTypeAccordionInBestellWizard
 */
export interface ProductTypeAccordionInBestellWizard {
  /**
   *
   * @type {string}
   * @memberof ProductTypeAccordionInBestellWizard
   */
  title: string;
  /**
   *
   * @type {string}
   * @memberof ProductTypeAccordionInBestellWizard
   */
  description: string;
  /**
   *
   * @type {number}
   * @memberof ProductTypeAccordionInBestellWizard
   */
  order: number;
}

/**
 * Check if a given object implements the ProductTypeAccordionInBestellWizard interface.
 */
export function instanceOfProductTypeAccordionInBestellWizard(
  value: object,
): value is ProductTypeAccordionInBestellWizard {
  if (!("title" in value) || value["title"] === undefined) return false;
  if (!("description" in value) || value["description"] === undefined)
    return false;
  if (!("order" in value) || value["order"] === undefined) return false;
  return true;
}

export function ProductTypeAccordionInBestellWizardFromJSON(
  json: any,
): ProductTypeAccordionInBestellWizard {
  return ProductTypeAccordionInBestellWizardFromJSONTyped(json, false);
}

export function ProductTypeAccordionInBestellWizardFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ProductTypeAccordionInBestellWizard {
  if (json == null) {
    return json;
  }
  return {
    title: json["title"],
    description: json["description"],
    order: json["order"],
  };
}

export function ProductTypeAccordionInBestellWizardToJSON(
  json: any,
): ProductTypeAccordionInBestellWizard {
  return ProductTypeAccordionInBestellWizardToJSONTyped(json, false);
}

export function ProductTypeAccordionInBestellWizardToJSONTyped(
  value?: ProductTypeAccordionInBestellWizard | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    title: value["title"],
    description: value["description"],
    order: value["order"],
  };
}